fix: translate_package_lock handles semver ranges
We have to do a second lookup through the file to get the actual version
of the package that satisfied the dependency.

Also add the ability to patch npm packages as we fetch them,
since there's no way to use the patch-package package in this setup
where npm packages are installed individually.