Fix setting log level in run_js_binary
The dictionary value of `env` was being masked over by a for loop,
resulting in a string being passed to `dicts.add`, which caused it to
error out.

Use a new variable name when iterating over the log level env values and
update an example to set the log_level so that this doesn't regress in
the future.