refactor: add way to suppress warnings for npm_translate_lock missing data
If a user doesn't start with lockfile updating, enabling it can be noisy. This enables it to be suppressed. There should also be a follow up pull request that aggregates the failures to a list and prints out a formatted data attribute that can be copy pasted so users don't have to go line by line.