fix: detect GNU tar on non-linux environments (#1084)
* fix: detect GNU tar on non-linux environments

I had a hard to debug failure which was due to the fact that on my
darwin machine I had gnutar on the $PATH as tar. Without the extra flag
"--delay-directory-restore" I would get failures due to permission
errors. This commit updates ensures we always add that flag if the
version output of tar reports it is GNU tar.

* Update npm_import.bzl

---------

Co-authored-by: Alex Eagle <alex@aspect.dev>