fix: `js_run_binary` doesn't accept `select()` (#2137)
I have a situation internally where I need to pass a different environment variable based on CI or remote development. Right now, `select` is not a `dict` so this fails. This is not an issue with skylibs `run_binary` or `bazel-lib`'s `run_binary`. By flipping this to a `|` merger, the issue goes away.

---------

Co-authored-by: Jason Bedard <jason+github@jbedard.ca>