fix: duplicate package versions with vendored and external (#2332)
Fix multiple copies of a package with the same name+version but from
different sources incorrectly overwriting each other. This fixes when a
package is from a registry and a file, there is still an issue when
there are 2 copies of a package but from different registries.

This is one thing found when originally adding support for pnpm v9 and
left as a `fail()` while adding a `print(WARNING)` to <v9. When
upgrading the root rules_js pnpm-lock.yaml to v9 this issue must.

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: yes
- Breaking change (forces users to change their own code or config): no
- Suggested release notes appear below: yes

Repository names for vendored packages such as tarballs may change
depending on the version protocol. The `{version}` in things like the
`@npm__{package}__{version}` repo name will now remain as the version
specified in package.json (the `file:` version) even if the resolved
package does contain a friendly version.

### Test plan

- Covered by existing test cases
- New test cases added