Fix fetching npm packages from git repositories
Aspect_rules_js supports fetching a git repository containing an npm
package. However, to follow that path, it is required that the URL
of the repository is prefixed with `git+`. However, this URL is then
also configured as the remote for the git repository. The `git+`
prefix is not supported by git and thus fetching fails with:

git: 'remote-git+https' ist kein Git-Befehl. Siehe 'git --help'.

To prevent this, the `git+`-prefix is now being removed, if it
is present before setting it as remote url.