fix: treat file: references in package store the same as pnpm (#2445)
This aligns rules_js with pnpm when it comes to `file:` references in
the package store, as well as bazel target throughout the various
store/link targets.

Now only `workspace:*` deps (which resolve to `link:path` versions) will
continue to use `@0.0.0` in the package store. This is different then
pnpm where local workspace deps are symlinks into source and do not
exist in the package store at all.

Ref #2300 which will depend on this change.

### Changes are visible to end-users: no

### Test plan

- Covered by existing test cases
- New test cases added