Prepare for fixing double-parsing (#2477)
Currently we parse the lockfile in both the extension and the repo rule
(so we can generate all the bzl/BUILD file). The plan to fix that is to
move the file generation into the module extension and reuse the parse
state. To do that, we must decouple the generation from the side effects
(dumping files to disk).

While here, I changed the args to `generate` to enforce it becomes pure.

---

### Changes are visible to end-users: no

### Test plan
- Covered by existing test cases