fix: fetching npm packages from git repositories (#2403)
Aspect_rules_js supports fetching a git repository containing an npm
package. However, to follow that path, it is required that the URL of
the repository is prefixed with `git+`. However, this URL is then also
configured as the remote for the git repository. The `git+` prefix is
not supported by git and thus fetching fails with:

git: 'remote-git+https' ist kein Git-Befehl. Siehe 'git --help'.

To prevent this, the `git+`-prefix is now being removed, if it is
present before setting it as remote url.formation ends up in the git
log.

---

### Changes are visible to end-users: no

### Test plan

- Manual testing; please provide instructions so we can reproduce:

Install an NPM package from a git repository. The pnpm-lock.yaml may
look like this:

```
lockfileVersion: '6.0'

settings:
  autoInstallPeers: true
  excludeLinksFromLockfile: false

dependencies:
  highlight.js:
    specifier: ^11.11.1
    version: 11.11.1
  highlightjs-ttcn3:
    specifier: https://gitea.osmocom.org/ttcn3/highlightjs-ttcn3.git#6daccff309fca1e7561a43984d42fa4f829ce06d
    version: gitea.osmocom.org/ttcn3/highlightjs-ttcn3.git#6daccff309fca1e7561a43984d42fa4f829ce06d

packages:
  /highlight.js@11.11.1:
    resolution: {integrity: sha512-Xwwo44whKBVCYoliBQwaPvtd/2tYFkRQtXDWj1nackaV2JPXx3L0+Jvd8/qCJ2p+ML0/XVkJ2q+Mr+UVdpJK5w==}
    engines: {node: '>=12.0.0'}
    dev: false

  gitea.osmocom.org/ttcn3/highlightjs-ttcn3.git#6daccff309fca1e7561a43984d42fa4f829ce06d:
    resolution:
      type: git
      repo: git+https://gitea.osmocom.org/ttcn3/highlightjs-ttcn3.git
      commit: 6daccff309fca1e7561a43984d42fa4f829ce06d
    name: highlightjs-ttcn3
    version: 0.0.1
    dependencies:
      highlight.js: 11.11.1
    dev: false
```